cnn-lab/requirements.txt
# Deep Learning Framework
# PyTorch는 CUDA 버전에 따라 별도 설치 필요 (README.md 참고)
# pip install torch==2.4.0 torchvision==0.19.0 torchaudio==2.4.0 --index-url https://download.pytorch.org/whl/cu118

# --- Model & Vision ---
timm==0.9.11
facenet-pytorch==2.5.3
grad-cam==1.5.3

# --- Data Processing ---
numpy==1.26.4
pandas==2.2.2
Pillow==11.2.1
opencv-python==4.10.0.84

# --- Visualization ---
matplotlib==3.10.0
seaborn==0.13.2

# --- Metrics ---
scikit-learn==1.4.0

# --- Utilities ---
PyYAML==6.0.2
tqdm==4.67.1
